requirements.txt
aprslib
fastapi
uvicorn
psycopg2-binary